some functions of 'table.py' were moved to separate file
